refactored about binding adapter
